Tilføjet PersonResource-klassen og REST API-metoderne.
Denne commit tilføjer PersonResource-klassen, som indeholder REST API-metoder til at få adgang til og manipulere persondata. Klassen bruger PersonFacaden til at håndtere selve dataadgangen.

Følgende REST API-metoder er implementeret:

- GET /personer: Returnerer en liste over alle personer
- GET /persons/{id}: Returnerer en bestemt person efter ID
- POST /personer: Tilføjer en ny person
- PUT /persons/{id}: Redigerer en bestemt person efter ID
- SLET /persons/{id}: Sletter en bestemt person efter ID

Alle metoder returnerer JSON-formaterede data.

+++ src/main/java/rest/PersonResource.java
@@ -21,25 +21,45 @@ public class PersonResource {
 
     private final PersonFacade facade = new PersonFacade();
 
+    /**
+     * Henter en liste over alle Person-entiteter i databasen.
+     * @return En liste af PersonDTO'er.
+     */
     @GET
     public Response getAllPersons() {
         List<PersonDTO> personDTOs = facade.getAllPersons();
         return Response.ok(personDTOs).build();
     }
 
+    /**
+     * Henter en Person-entitet ud fra dens id.
+     * @param id Id'en på den ønskede Person-entitet.
+     * @return En PersonDTO.
+     */
     @GET
     @Path("{id}")
-    public Response getPersonById(@PathParam("id") long id) {
+    public Response getPersonById(@PathParam("id") int id) {
         PersonDTO personDTO = facade.getPersonById(id);
         return Response.ok(personDTO).build();
     }
 
+    /**
+     * Tilføjer en Person-entitet til databasen.
+     * @param personDTO PersonDTO'en der skal tilføjes.
+     * @return Den tilføjede PersonDTO.
+     */
     @POST
     public Response addPerson(PersonDTO personDTO) {
         PersonDTO addedPersonDTO = facade.addPerson(personDTO);
         return Response.ok(addedPersonDTO).build();
     }
 
+    /**
+     * Opdaterer en Person-entitet i databasen.
+     * @param id Id'en på den Person-entitet der skal opdateres.
+     * @param personDTO PersonDTO'en der skal opdateres.
+     * @return Den opdaterede PersonDTO.
+     */
     @PUT
     @Path("{id}")
     public Response editPerson(@PathParam("id") int id, PersonDTO personDTO) {
@@ -48,9 +68,15 @@ public class PersonResource {
         return Response.ok(editedPersonDTO).build();
     }
 
+    /**
+     * Sletter en Person-entitet fra databasen.
+     * @param id Id'en på den Person-entitet der skal slettes.
+     * @return En Response med status 200 OK hvis Person-entiteten blev slettet
+     * eller 404 Not Found hvis den ikke blev fundet i databasen.
+     */
     @DELETE
     @Path("{id}")
-    public Response deletePerson(@PathParam("id") long id) {
+    public Response deletePerson(@PathParam("id") int id) {
         facade.deletePerson(id);
         return Response.ok().build();
     }
